Safar-Ticket-Backend/requirements.txt
Django>=4.2
djangorestframework
PyMySQL
PyJWT
django-redis
cryptography>=3.0
django-cors-headers
elasticsearch==8.11.1